fixed a bug where the favourite button in plant fragments was not correctly updated due to new database async listeners

+++ app/src/main/java/com/example/urbotanist/database/DatabaseRetriever.java
@@ -47,7 +47,9 @@ public class DatabaseRetriever {
       public void execute(@NonNull Realm realm) {
         FavouritePlant favouriteToDelete = realm.where(FavouritePlant.class)
             .equalTo("plantId", plantId).findFirst();
-        favouriteToDelete.deleteFromRealm();
+        if(favouriteToDelete != null) {
+          favouriteToDelete.deleteFromRealm();
+        }
       }
     });
   }
